+++ HISTORY.md
@@ -4,6 +4,9 @@
 
 ## [unreleased/patches]
 
+## [0.4.0] - 2026-01-05 | Player plugin settings implemented
+
+- feat: player plugin settings implemented
 - fix: botLang is null when Discord Connect is not found [v0.3.1]
 
 ## [0.3.0] - 2025-12-28 | Changed Consumer to Callback